api/requirements.txt
amqp==5.3.1
asgiref==3.10.0
attrs==25.4.0
autobahn==25.10.2
Automat==25.4.16
billiard==4.2.1
celery==5.5.3
certifi==2025.6.15
cffi==2.0.0
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudinary==1.44.1
constantly==23.10.4
cryptography==46.0.3
daphne==4.2.1
Django==5.2.3
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django-filter==25.1
django-storages==1.14.6
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
et_xmlfile==2.0.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
kombu==5.5.4
MarkupSafe==3.0.3
msgpack==1.1.2
openpyxl==3.1.5
packaging==25.0
pillow==11.2.1
prompt_toolkit==3.0.51
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyJWT==2.9.0
pyOpenSSL==25.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-http-client==3.3.7
redis==6.4.0
requests==2.32.4
resend==2.19.0
sendgrid==6.12.5
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
Twisted==25.5.0
txaio==25.9.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.4
zope.interface==8.1
